+++ lib/guardian_searcher/options.rb
@@ -1,16 +1,21 @@
 # frozen_string_literal: true
 
 module GuardianSearcher
+  # the class maps the options passed to the ones needed
+  # by the Guardian API for searching
   class Options < Hash
-    private attr_accessor :options
-
     def method_missing(method_name, *args, &blk)
       return options.[](method_name, &blk) if @options.key?(method_name)
 
       super(method_name, *args, &blk)
     end
 
+    def respond_to_missing?(method_name, *args)
+      @options.key?(method_name) || super(method_name, *args)
+    end
+
     def initialize(options)
+      super
       raise OptionsNotHashError unless options.is_a?(Hash)
 
       @options = options
@@ -38,5 +43,9 @@ module GuardianSearcher
         page: "page"
       }[key]
     end
+
+    private
+
+    attr_accessor :options
   end
 end