+++ lib/guardian_searcher/search_result.rb
@@ -21,7 +21,27 @@ module GuardianSearcher
       @start = start
     end
 
-    def self.parse_results(body: nil)
+    def self.parse_with_codes(response: nil)
+      raise GuardianSearcher::GuardianSearcherUndefinedResponse unless response
+
+      case response.status
+      when 200
+        parse_results(body: response.body)
+      when 400
+        message = JSON.parse(response.body)["message"]
+        raise GuardianSearcher::GuardianBadRequestError, message
+      when 401
+        message = JSON.parse(response.body)["message"]
+        raise GuardianSearcher::GuardianUnauthorizedError, message
+      when 500
+        message = JSON.parse(response.body)["message"]
+        raise GuardianSearcher::GuardianInternalServerError, message
+      else
+        raise GuardianSearcher::GuardianUnknownError, "Unknown error, check Faraday response"
+      end
+    end
+
+    def self.parse_results(body:)
       return unless body
 
       body = JSON.parse(body)